refactor: AWS, S3 설정 모듈화, File, AWS 패키지 분리, Path 인터페이스 추가
