+++ src/main/java/com/kdt/team04/common/file/FileStorage.java
@@ -1,10 +1,10 @@
-package com.kdt.team04.common.file.service;
+package com.kdt.team04.common.file;
 
 import org.springframework.core.io.Resource;
 
 public interface FileStorage {
 
-	String uploadByPath(Resource resource, String path);
+	String uploadByPath(Resource resource, Path path);
 
 	void uploadByKey(Resource resource, String key);
