+++ src/main/java/com/kdt/team04/common/file/Path.java
@@ -1,0 +1,5 @@
+package com.kdt.team04.common.file;
+
+public interface Path {
+	String getPath();
+}
